Move checkValidations
