+++ src/applications/appeals/995/tests/validations/validations.unit.spec.js
@@ -2,11 +2,25 @@ import { expect } from 'chai';
 import sinon from 'sinon';
 
 import {
+  checkValidations,
   requireRatedDisability,
   contactInfoValidation,
   missingPrimaryPhone,
 } from '../../validations';
+import { missingIssueName } from '../../validations/issues';
+
 import { errorMessages, SELECTED, PRIMARY_PHONE } from '../../constants';
+
+describe('checkValidations', () => {
+  it('should return error messages', () => {
+    expect(checkValidations([missingIssueName], '')).to.deep.equal([
+      errorMessages.missingIssue, // simple validation function
+    ]);
+    expect(
+      checkValidations([missingIssueName, missingIssueName], ''),
+    ).to.deep.equal([errorMessages.missingIssue, errorMessages.missingIssue]);
+  });
+});
 
 describe('requireRatedDisability', () => {
   it('should show an error if no disabilities are selected', () => {
