+++ src/applications/appeals/995/validations/issues.js
@@ -5,22 +5,6 @@ import {
 } from '../content/contestableIssues';
 import { errorMessages, MAX_LENGTH } from '../constants';
 
-/**
- * Check validations for Custom pages
- * @param {Function[]} validations - array of validation functions
- * @param {*} data - field data passed to the validation function
- * @param {*} fullData - full and appStateData passed to validation function
- * @returns {String[]} - error messages
- */
-export const checkValidations = (validations, data = {}, fullData = {}) => {
-  const errors = { errorMessages: [] };
-  errors.addError = message => errors.errorMessages.push(message);
-  validations.map(validation =>
-    validation(errors, data, fullData, null, null, null, fullData),
-  );
-  return errors.errorMessages;
-};
-
 export const selectionRequired = (
   errors,
   _fieldData,
